move getDomain function
